moving ActorTest-style tests from common to pslogin
